+++ client/src/api/regions.js
@@ -11,3 +11,10 @@ export function leaveRegion (regionId) {
 export function masterUpdate (regionId) {
   return patch(`/region/${regionId}/masterupdate`)
 }
+
+export function setRegionOptions (regionId, enableReportButton, enableMediationButton) {
+  return post(`/region/${regionId}/options`, {
+    enableReportButton: enableReportButton,
+    enableMediationButton: enableMediationButton,
+  })
+}